Fix PDU filter in mixin
`values_list` by default returns a tuple, even if there's only one field in each entry.
As such the check lower down didn't work, as it was looking for the PDU string in a list of tuples.

The fix is to pass flat=True which makes it return a list of values not tuples.

Before this fix a standard user (whatever role but not a superuser or RCPCH team) would get a full-screen 403 accessing either the list of patients or users.