Set session data in signal
At the moment the tests aren't actually logged in, hence the check for either a 200 or a 302 code. This didn't matter in our current tests since we were inspecting the queryset directly but does matter for future tests that send POST requests.

Swapping the sessionid in client cookies didn't work.

I've removed that entirely and changed the session data to be added in the user_logged_in signal. That way we don't need the session swapping test code at all and we are actually logged in during tests.