Embed build info into our image
The previous attempt in #208 worked locally but not in the container because .git is ignored in our .dockerignore file. I've changed the whole process to write a cache file in CI which is burnt into the image.

When running locally there is no cache file but for simplicity it calls the same bash script to get the version info. In all cases the build info is then cached in a global variable to avoid extra i/o on every request.

Unfortunately creating the build info does need to be in a separate script that can be run from the host as the dockerignore means we will never have git information in the container. That means we can't add creating the build info to the Dockerfile.