Fixes #207. Periodically get a new Azure DB token
The tokens live for a day but 15 minutes seems like a reasonable amount of time to avoid doing it too much. From what I can tell the token is cached somewhere within Azure Container Apps anyway. I guess *really* we should drive this period from the expiry time on the token itself but nothing else in the app parses it yet so this seems a reasonable compromise.