Add an http access log
Seemed simplest to do this in gunicorn as the Django logging configuration is very complicated and it's Friday and I can't cope with that before a weekend.

The logging format is the default format but I have replaced the host IP logging with X-Forwarded-For as otherwise it will always be the same IP address of the Azure ingress machine.