Fix error handling for rows missing both an nhs number and a urn.
We already had errors in the form but now I've rearranged the code we weren't saving them - just the special case error that was in `process_rows_for_patient`.

I've moved the guard from the special case into `save_patient_and_transfer` so that we have an opportunity to gather the existing errors first.

This means we can remove the redundant `__all__` error in favour of the existing errors on both fields that already existed from the form.