Make test_kpi 5 and 6 pass reliably
They started flaking when we crossed over from audit year 24-25 to 25-26 (1st of April). I think this is because the PatientFactory uses random dates of diagnosis and there weren't many days left in the audit year before and now there are lots, so the tests are more likely to flake.

The fix is to pin down diagnosis_date on even eligible patients (as we do elsewhere in the same tests).