Prepatory csv upload code for Celery
Split creating the submission, processing the file and tidying up old submissions into separate methods so they can be called in different places in the future (http route and Celery tasks).

Add the VisitActivity entry as soon as we've created the submission so we still have a record of the upload even if it crashes.

Delete patients from old submissions after we have successfully processed the new one to make it easier to revert back in case of a crash processing the new file.