Add test setting to crash on startup
This will help us investigate #1089. I have enabled the setting on staging but not on live.

If the default Azure startup and readiness probes are working, this should mean this PR fails to deploy on staging but does deploy on live. Both instances should remain accessible throughout.

If this reproduces the issue from #1089 then staging will become unresponsive. Live should be unaffected as the setting is not enabled.