Fix accessing direct links to function based views when initially logged out.
login_required needs to be used as a wrapper decorator otherwise the result is discarded. Luckily check_otp will fail too if the user is logged out so anonymous users were still forced to log in, but we ate the `next` query parameter so they didn't end up at the page they wanted.